+++ packages/react-property-selectors/src/hooks/combobox/index.ts
@@ -1,2 +1,0 @@
-export * from "./use-combobox-property-selector";
-export * from "./use-image-combobox-property-selector";