+++ packages/react-property-selectors/src/index.tsx
@@ -5,7 +5,6 @@ export * from "./components/radio-group";
 export * from "./components/textbox";
 // Hooks
 export * from "./hooks/amount/index";
-export * from "./hooks/checkbox/index";
-export * from "./hooks/combobox/index";
-export * from "./hooks/radio-group";
 export * from "./hooks/textbox";
+export * from "./hooks/discrete";
+export * from "./hooks/properties-selector";