+++ packages/react-property-selectors/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/promaster-sdk/property/compare/@promaster-sdk%2Freact-property-selectors@7.0.0...master)
 
+## Changed
+
+- Added generic DiscretePropertySelector hooks and removed specific hooks for Checkbox, Combobox, ImageCombo, RadioGroup.
+- All discrete selectors now share a single hook.
+- Move hook for usePropertiesSelector from package react-properties-selector to package react-property-selectores. Now all hooks are in the same package and the react-properties-selector package is deprecated.
+- You can pass any items to the hooks. It now uses callbacks to extract data from your items. This means you can use any extra data you want in the items and use that during rendering.
+
 ## [v7.0.0](https://github.com/promaster-sdk/property/compare/@promaster-sdk%2Freact-property-selectors@6.0.0...@promaster-sdk%2Freact-property-selectors@7.0.0) - 2020-10-15
 
 ### Added